Standardized Formatting
Expresso was written by multiple people (I think) and different source files had different formatting.
Applied my clangformat/visual studio formatting information in terms of indentation and brackets.

as bad as forcing my own formatting requirements on third party code is I can't read it when it looks like
for(...)
for(...){
if(...)
 while(...);
}for(...)

instead of
for(...){
    for(...){
        if(...)
            while(...)
    }
}